:label: climbing 타입 리팩토링 (#266)
