+++ src/types/apis/climbing.ts
@@ -1,64 +1,25 @@
-import type { BookInfo } from '@src/types/apis/book.d';
-import { ClimbingStatus } from '@src/constants/constants';
+import { Climbing, ClimbingMember } from '@src/types/climbing';
+import { EmojiTypeType } from '@src/constants/constants';
 
-export type book = {
-  author: string;
-  cover: string;
-  description: string;
-  isbn13: string;
-  itemPage: string;
-  pubDate: string;
-  publisher: string;
-  title: string;
+export type patchClimbingChannelReq = Pick<
+  Climbing,
+  'name' | 'description' | 'startDate' | 'endDate'
+> & {};
+
+export type getClimbingChannelMembersRes = {
+  climbingMemberList: ClimbingMember[];
 };
 
-export type Climbing = {
-  climbingId: number;
-  status: ClimbingStatus;
-  name: string;
-  startDate: string;
-  endDate: string;
-  description: string;
-  memberCount: number;
-  isJoined: boolean;
-  isOwner: boolean;
-  bookInfo: book;
+export type patchClimbingMemoReq = {
+  memo: string | null;
 };
 
-// member에서 가져올 것
-type EmojiMember = {
-  memberId: number;
-  nickname: string;
-  profileImg: string | null;
-  level: number;
-  mountain: string;
-};
-
-export type ReviewEmoji = {
-  emoji: string;
-  reviewEmojiMemberList: EmojiMember[];
-};
-
-export type ReviewEmojiResponse = {
-  reviewEmojiList: ReviewEmoji[];
-};
-
-type ClimbingMemberReview = {
-  memberId: number;
-  profileImg: string;
-  nickname: string;
-  star: number;
-  reviewId: number;
-  content: string;
-  reviewEmojiList: { emoji: string; emojiCount: number }[];
-};
-
-export type ClimbingResponse = {
-  hasShared: boolean;
-  isShareable: boolean;
-  bookInfo: BookInfo;
-  content?: string;
-  reviewId?: number;
-  star?: number;
-  ClimbingMemberReviewList?: ClimbingMemberReview[];
+export type getClimbingReviewEmojiRes = {
+  reviewEmojiList: {
+    emoji: EmojiTypeType;
+    reviewEmojiMemberList: Pick<
+      ClimbingMember,
+      'memberId' | 'nickname' | 'profileImg' | 'level' | 'mountain'
+    >[];
+  };
 };
